Sprites can change appearance
We have a skript now that can change a Sprites appearance. One of the sprites should be smol for this and the other one should be big and contain text that you can read or something. If you click it, you can read i and click again to put it away again.